[Refactor] - Moves language settings classes into localiztion_ui module
