+++ access_card/src/main/res/values-es/strings.xml
@@ -10,5 +10,5 @@
     <string name="accessMemberCardLabel">Tarjeta de acceso para miembros</string>
     <string name="expires">Vence: %s</string>
     <string name="zipCodeHint">Ingrese su código postal…</string>
-    <string name="memberIdHint">Ingrese la Identificación del Miembro...</string>
+    <string name="memberIdHint">Ingrese la Identificación del Miembro…</string>
 </resources>
