+++ access_card/src/main/res/values/strings.xml
@@ -1,8 +1,4 @@
 <resources>
-    <string name="english" translatable="false">English</string>
-    <string name="spanish" translatable="false">Español</string>
-    <string name="chinese" translatable="false">中文</string>
-
     <string name="becomeAMember">Become a Member</string>
     <string name="signIn">Sign In</string>
     <string name="homeZipCode">Zip Code</string>
@@ -12,6 +8,6 @@
     <string name="memberCard">Member Card</string>
     <string name="accessMemberCardLabel">Access Member Card</string>
     <string name="expires">Expires: %s</string>
-    <string name="zipCodeHint">Enter your home zip code...</string>
-    <string name="memberIdHint">Enter your Member ID...</string>
+    <string name="zipCodeHint">Enter your home zip code…</string>
+    <string name="memberIdHint">Enter your Member ID…</string>
 </resources>
