+++ localization_ui/src/main/java/artic/edu/localization/ui/LanguageSettingsFragment.kt
@@ -1,4 +1,4 @@
-package edu.artic.info
+package artic.edu.localization.ui
 
 
 import android.os.Bundle
