+++ localization_ui/src/main/java/artic/edu/localization/ui/LanguageSettingsViewModel.kt
@@ -1,4 +1,4 @@
-package edu.artic.info
+package artic.edu.localization.ui
 
 import com.fuzz.rx.disposedBy
 import edu.artic.analytics.AnalyticsAction
