+++ localization_ui/src/main/res/color/info_menu_color_list.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="@color/brownishOrange" android:state_checked="true" />
+    <item android:color="@color/gray" />
+</selector>
