+++ localization_ui/src/main/res/color/radiobutton_text_color_list.xml
