+++ localization_ui/src/main/res/drawable/language_radio_button_background.xml
