+++ localization_ui/src/main/res/values/strings.xml
@@ -1,0 +1,9 @@
+<resources>
+    <string name="languageSettings">Language Settings</string>
+    <string name="languageSettingsBody">Some content may not be available in your selected language.</string>
+    <string name="chooseYourLanguage">Please Choose Your Preferred Language</string>
+    <string name="english" translatable="false">English</string>
+    <string name="spanish" translatable="false">Español</string>
+    <string name="chinese" translatable="false">中文</string>
+
+</resources>
