+++ localization_ui/src/main/res/values/styles.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="LanguageButton">
+        <item name="android:fontFamily">@font/ideal_sans_medium</item>
+        <item name="android:textAllCaps">false</item>
+        <item name="android:textColor">@color/radiobutton_text_color_list</item>
+        <item name="android:textSize">15sp</item>
+        <item name="android:gravity">center</item>
+        <item name="android:textAlignment">center</item>
+        <item name="android:minHeight">50dp</item>
+    </style>
+</resources>
