+++ settings.gradle
@@ -12,6 +12,7 @@ include ':adapter',
         ':image',
         ':info',
         ':localization',
+        ':localization_ui',
         ':location',
         ':location_ui',
         ':map',
